first shot at moving linebuffer into all the modules
