+++ src/Core/Aetherling/Simulator/Arrays.hs
@@ -135,85 +135,63 @@ simhlMunchArray op t (value:values) =
 simhlMunchArray op _ _ =
     error ("Aetherling internal error: broken munch for " ++ show op)
 
--- | Line buffer simulator implementation.
---
--- THIS FUNCTION IS OUT-OF-DATE DUE TO The Line Buffer Manifesto.
---
--- For now I'm only simulating 1D and 2D linebuffers. There's some
--- important restrictions on pixels-per-clock UPDATE THIS IF/WHEN
--- THAT'S NO LONGER TRUE (including in simhlPre).  pixels-per-clock in
--- the 2D case is passed as a list of [width, height]. For now, height
--- must be 1, and width (for 1D + 2D) must divide both the image width
--- and 1 minus the window width. Justification: never "cross" a
--- no-output with-output boundary in a single cycle.
---
--- Implement by dumping all input data into a Haskell array
--- (representing the image being streamed in), making a list of output
--- windows from that array, and finally splitting that output list
--- into the [[ValueType]] output expected.
-simhlLineBuffer :: [Int] -> [Int] -> [Int] -> TokenType -> [[ValueType]] ->
-                BoundaryConditions -> [[ValueType]]
-simhlLineBuffer [1, pixW] [wH, wW] [iH, iW] t [inStr] bc =
-    let
-      -- Part 1: Make the 2D array.
-      -- Note that our array has indicies swapped compared to
-      -- typical width-height representation. Justification:
-      -- lexicographical order everywhere.
-      bounds = ((0,0), (iH-1, iW-1))
-      getValues = simhlSerializeArray (T_Array 1 (T_Array pixW t))
-      values = concat [getValues a | a <- inStr] ++ repeat V_Unit
-      the_array = listArray bounds values
-      
-      -- Split it up into windows
-      -- Make window with pixel (y,x) at lower-right.
-      mkWindow y x = V_Array [
-          V_Array [ the_array ! (y',x') | x' <- [x-wW+1..x] ]
-          | y' <- [y-wH+1..y]
-        ]
-      windows = [mkWindow y x | y <- [wH-1..iH-1], x <- [wW-1..iW-1]]
-      -- Munch the windows and pack them into a stream of output
-      -- arrays-of-arrays-of-arrays-of-arrays My understanding of the
-      -- 4D output is that the inner 2 dims correspond to the window
-      -- dimensions, and the outer 2 dims correspond to the pixel
-      -- inputs. So (i,j,k,l) would be the (row k, col l) pixel of the
-      -- window with the pixel inputted at position (row i, col j) at
-      -- its lower-right.
-      pack :: [ValueType] -> [ValueType]
-      pack w | length w <= pixW = [V_Array [ V_Array [a | a <- w]]]
-      pack w =
-        let
-          (these_windows, later_windows) = splitAt pixW w
-          this_array = V_Array [ V_Array [a | a <- these_windows]]
-          later_arrays = pack later_windows
-        in
-          this_array:later_arrays
-    in
-      [pack windows]
-simhlLineBuffer [pixW] [wW] [iW] t [inStr] bc =
-    let
-      bounds = (0, iW-1)
-      getValues = simhlSerializeArray (T_Array pixW t)
-      values = concat [getValues a | a <- inStr] ++ repeat V_Unit
-      the_array = listArray bounds values
+simhlLineBuffer :: LineBufferData -> [[ValueType]] -> [[ValueType]]
+simhlLineBuffer lbData [inStr] =
+  let
+    (yPerClk, xPerClk) = lbPxPerClk lbData
+    (originY, originX) = lbOrigin lbData
+    (windowY, windowX) = lbWindow lbData
+    (strideY, strideX) = lbStride lbData
+    (imgY, imgX) = lbImage lbData
 
-      mkWindow x = V_Array [ the_array ! x' | x' <- [x-wW+1..x] ]
+    -- We'll simulate it just by filling up an entire array with pixels
+    -- from the input, then carve up the array into outputs.
+    -- Remember that everything's y, x so it's lexicographical order!
+    bounds = ((0,0), (imgY-1, imgX-1))
 
-      windows = [mkWindow x | x <- [wW-1..iW-1] ]
+    -- Convert input array to stream of xPerClk pixels (assumed yPerClk = 1).
+    unpackInput :: ValueType -> [ValueType]
+    unpackInput V_Unit = replicate xPerClk V_Unit
+    unpackInput (V_Array [V_Array a]) = a
+    unpackInput _ = error "Aetherling internal error: expected 1-by-x array \
+      \input in line buffer."
 
-      pack :: [ValueType] -> [ValueType]
-      pack w | length w <= pixW = [V_Array [a | a <- w]]
-      pack w =
-        let
-          (these_windows, later_windows) = splitAt pixW w
-          this_array = V_Array [a | a <- these_windows]
-          later_arrays = pack later_windows
-        in
-          this_array:later_arrays
-    in
-      [pack windows]
-simhlLineBuffer _ _ _ _ _ _ =
-    error "Aetherling intenal error: Unexpected LineBuffer parameters"
+    -- Use the fact that pixels are in lexicographical order to make
+    -- the array. If the input was too short, fill the end with
+    -- V_Unit.
+    values = concat (map unpackInput inStr) ++ repeat V_Unit
+    pixelArray = listArray bounds values
 
+    -- Create the output windows using this here helper function.
+    -- (y, x) is the upper-left of the window.
+    mkWindow y x = V_Array [
+        V_Array [
+          if y' < 0 || y' >= imgY || x' < 0 || x' >= imgX then V_Unit
+          else pixelArray ! (y',x')
+          | x' <- [x..x+windowX-1]
+        ] | y' <- [y..y+windowY-1]
+      ]
+    windowYs = [originY,originY+strideY..originY+imgY-1]
+    windowXs = [originX,originX+strideX..originX+imgX-1]
+    windows = [mkWindow y x | y <- windowYs, x <- windowXs]
+
+    -- Now, we need to calculate the parallelism to know how many
+    -- windows we emit per (logical) cycle. Then munch the arrays
+    -- and pack them into the 4D output arrays.
+    parallelism = getParallelism lbData
+    pack :: [ValueType] -> [ValueType]
+    pack [] = []
+    pack windows =
+      let
+        (theseWindows, laterWindows) = splitAt parallelism windows
+        thisOut = V_Array theseWindows
+        laterOut = pack laterWindows
+      in
+        thisOut:laterOut
+  in
+    [pack windows]
+simhlLineBuffer _ _ = error "Aetherling internal error: expected 1 \
+  \input stream for line buffer."
 
 -- | Preprocessor pass implementation for SequenceArrayRepack.
 simhlPreRepack :: [Op] -> [Maybe Int] -> SimhlPreState
@@ -263,55 +241,20 @@ simhlPreReshape _ _ _ =
 
 -- | Preprocessor pass implementation for LineBuffer.
 --
--- Check that the LineBuffer conforms to the restrictions commented on
--- in simhlLineBuffer.
-simhlPreLB :: [Op] -> [Maybe Int] -> SimhlPreState
-           -> ([Maybe Int], SimhlPreState)
-simhlPreLB opStack@(LineBuffer [pixW] [wW] [iW] t _:_) [inStrLen] inState =
-    if (wW-1) `mod` pixW /= 0 || iW `mod` pixW /= 0 || any (<=0) [pixW, wW, iW] then
-      error("1D LineBuffer requires all positive parameters and the width of \
-            \pxPerClock to divide both the image width and one minus the window \
-            \width, at\n"
-            ++ (simhlFormatOpStack opStack)
-           )
-    else
-      let
-        strLen = div (iW-wW+1) pixW
-        expectedInStrLen = div iW pixW
-        just Nothing = expectedInStrLen
-        just (Just i) = i
-        warning' =
-          if expectedInStrLen /= just inStrLen then
-            Just "Unexpected input stream length"
-          else
-            Nothing
-      in
-        simhlPreResult opStack [Just strLen] warning' inState
+-- Part of the simulator preprocessor implementation for the line
+-- buffer. The line buffer always produces the same output stream length.
+-- If the input stream is longer or shorter than expected, return an
+-- Just String warning, otherwise no warning.
+simhlPreLB (LineBuffer lbData) [Nothing] =
+  (Nothing, [Just $ pSeqLen $ head $ outPorts lbData])
+simhlPreLB (LineBuffer lbData) [Just length] =
+  let
+    expected_length = pSeqLen $ head $ outPorts lbData
+    warning =
+      if expected_length == length then Nothing
+      else Just $ "Stream length (" ++ show length
+                  ++ ") not as expected (" ++ show expected_length ++ ")."
+  in
+    (warning, [Just $ pSeqLen $ head $ outPorts lbData])
 
-simhlPreLB opStack@(LineBuffer [1, pixW] [wH, wW] [iH, iW] t _:_) [inStrLen] inState =
-    if (wW-1) `mod` pixW /= 0 || iW `mod` pixW /= 0 || any (<=0) [pixW, wH, wW, iH, iW] then
-      error("2D LineBuffer requires all positive parameters and the width of \
-            \pixPerClock to divide both the image width and one minus the window \
-            \width, at\n"
-            ++ (simhlFormatOpStack opStack)
-      )
-    else
-      let
-        strLen = div ((iH-wH+1)*(iW-wW+1)) pixW
-        expectedInStrLen = div (iH*iW) pixW
-        just Nothing = expectedInStrLen
-        just (Just i) = i
-        warning' =
-          if expectedInStrLen /= just inStrLen then
-            Just "Unexpected input stream length"
-          else
-            Nothing
-      in
-        simhlPreResult opStack [Just strLen] warning' inState
-
-simhlPreLB opStack@(LineBuffer pix w i t _:_) _ _ =
-    error("Only support 1D linebuffers, and 2D linebuffers \
-          \with pixel window of form [1,w], at\n"
-       ++ (simhlFormatOpStack opStack)
-    )
 simhlPreLB _ _ _ = error "Aetherling internal error: expected LineBuffer."
